Use context to render table with EUI components
