+++ apps/rule-manager/client/package.json
@@ -21,6 +21,7 @@
 		"@types/jest": "^29.5.1",
 		"@types/react": "~17.0.53",
 		"@types/react-dom": "~17.0.19",
+		"@types/react-window-infinite-loader": "^1.0.6",
 		"@vitejs/plugin-react-swc": "^3.0.0",
 		"jest": "^29.5.0",
 		"sass": "~1.32.13",
